Make phone number configurable and add asterisk configuration
The phone number that will be displayed on the web page must be
specified as a flag when starting the vt-server.

Contains the asterisk configuration for use in a local network with two
example users `user_a` and `user_b`. The asterisk configuration uses
custom environment variables, which are explained in
`asterisk/extensions.conf`.